+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Define new TTP::Path::suffix() function
     - Make sure filenames read from the filesystem are utf-8 decoded
     - Define new audio.pl list verb
+    - Add and honor 'withPrefix' and 'withEol' options to msgOut()
 
 ### 4.18.3
 ### 4.18.2